+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Please document all notable changes to this project in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+
+* `getValueOrDefault` helper ([#125])
+
 ## [v0.1.6]
 
 ### Changed
@@ -64,6 +69,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Initial implementation
 
+[Unreleased]: https://github.com/projectsyn/commodore/compare/v0.1.6...HEAD
 [v0.1.0]: https://github.com/projectsyn/commodore/releases/tag/v0.1.0
 [v0.1.1]: https://github.com/projectsyn/commodore/releases/tag/v0.1.1
 [v0.1.2]: https://github.com/projectsyn/commodore/releases/tag/v0.1.2
@@ -82,3 +88,4 @@ Initial implementation
 [#100]: https://github.com/projectsyn/commodore/pull/100
 [#103]: https://github.com/projectsyn/commodore/pull/103
 [#106]: https://github.com/projectsyn/commodore/pull/106
+[#125]: https://github.com/projectsyn/commodore/pull/125