+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * `getValueOrDefault` helper ([#125])
 * `makeMergeable` helper ([#126])
 * `commodore component compile` to compile a single component ([#122])
+* labels to issue templates ([#134])
 
 ### Changed
 
@@ -100,3 +101,4 @@ Initial implementation
 [#125]: https://github.com/projectsyn/commodore/pull/125
 [#126]: https://github.com/projectsyn/commodore/pull/126
 [#130]: https://github.com/projectsyn/commodore/pull/130
+[#134]: https://github.com/projectsyn/commodore/pull/134