+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Option to explicitly set a component's display name ([#133])
 * labels to issue templates ([#134])
 * Vale Makefile target in component template ([#137])
+* Allow overriding Git user info for catalog commits ([#140])
 * Refactor tests to work with new setup-python ([#143])
 
 ### Changed
@@ -113,4 +114,5 @@ Initial implementation
 [#134]: https://github.com/projectsyn/commodore/pull/134
 [#136]: https://github.com/projectsyn/commodore/issues/136
 [#137]: https://github.com/projectsyn/commodore/pull/137
-[#137]: https://github.com/projectsyn/commodore/pull/143
+[#140]: https://github.com/projectsyn/commodore/pull/140
+[#143]: https://github.com/projectsyn/commodore/pull/143