+++ CHANGELOG.md
@@ -5,7 +5,8 @@ Please document all notable changes to this project in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [v0.2.0]
+
 ### Added
 
 * `getValueOrDefault` helper ([#125])
@@ -88,7 +89,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Initial implementation
 
-[Unreleased]: https://github.com/projectsyn/commodore/compare/v0.1.6...HEAD
+[Unreleased]: https://github.com/projectsyn/commodore/compare/v0.2.0...HEAD
 [v0.1.0]: https://github.com/projectsyn/commodore/releases/tag/v0.1.0
 [v0.1.1]: https://github.com/projectsyn/commodore/releases/tag/v0.1.1
 [v0.1.2]: https://github.com/projectsyn/commodore/releases/tag/v0.1.2
@@ -96,6 +97,7 @@ Initial implementation
 [v0.1.4]: https://github.com/projectsyn/commodore/releases/tag/v0.1.4
 [v0.1.5]: https://github.com/projectsyn/commodore/releases/tag/v0.1.5
 [v0.1.6]: https://github.com/projectsyn/commodore/releases/tag/v0.1.6
+[v0.2.0]: https://github.com/projectsyn/commodore/releases/tag/v0.2.0
 [#53]: https://github.com/projectsyn/commodore/pull/53
 [#58]: https://github.com/projectsyn/commodore/pull/58
 [#81]: https://github.com/projectsyn/commodore/pull/81